Add files via upload
In this project, the weights of the network are already trained and saved, and then the desired model is built based on the Lenet-5 network, and the weights are loaded into it, and the network is tested with the test data, and the network is tested using the dataset. Mnist and Fashion Mnist and Cifar-10 are implemented. At the end of the Test section, Accuracy (99.1) of the network is printed.